+++ core/src/Model/Enemies/Mouse.java
@@ -7,7 +7,7 @@ import com.dongbat.jbump.World;
 
 public class Mouse extends Enemy {
 
-    public Mouse(int x, int y, int height, int width, int speed, float health, float damage, World<Entity> world) {
+    public Mouse(float x, float y, float height, float width, float speed, float health, float damage, World<Entity> world) {
         super(x, y, height, width, speed, health, damage, world);
         setImagePath("Enemies/mouse-idle-down.png");
         addCollision();
